Add changeThreshold to OwnersContext
